fetched ride list

+++ lib/services/add_rides.dart
@@ -1,6 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 
-Future addRide(String name, int persons, String ride, double total) async {
+Future addRide(String name, int persons, String ride, int total) async {
   final docUser = FirebaseFirestore.instance.collection('Rides').doc();
 
   final json = {
